+++ src/refactor_nrepl/artifacts.clj
@@ -1,12 +1,18 @@
 (ns refactor-nrepl.artifacts
-  (:require [clojure.data.json :as json]
-            [clojure
-             [edn :as edn]
-             [string :as str]]
-            [clojure.java.io :as io]
-            [org.httpkit.client :as http]
-            [version-clj.core :as versions])
-  (:import java.util.zip.GZIPInputStream))
+  (:require
+   [cemerick.pomegranate :as pomegranate]
+   [clojure.data.json :as json]
+   [clojure.edn :as edn]
+   [clojure.java.io :as io]
+   [clojure.string :as str]
+   [clojure.tools.namespace.find :as find]
+   [org.httpkit.client :as http]
+   [refactor-nrepl.ns.slam.hound.regrow :as slamhound-regrow]
+   [refactor-nrepl.ns.slam.hound.search :as slamhound]
+   [version-clj.core :as versions])
+  (:import java.io.File
+           java.util.jar.JarFile
+           java.util.zip.GZIPInputStream))
 
 (def artifacts-file (str (System/getProperty "java.io.tmpdir")
                          "/refactor-nrepl-artifacts-cache"))
@@ -17,12 +23,15 @@
   (let [lm (.lastModified (io/file file))]
     (if (zero? lm) nil lm)))
 
-;;  structure here is {"prismatic/schem" ["0.1.1" "0.2.0" ...]}
-(defonce artifacts (atom (if (.exists (io/as-file artifacts-file))
-                           (->> artifacts-file slurp edn/read-string (into {}))
-                           {})
-                         :meta {:last-modified
-                                (get-last-modified-from-file artifacts-file)}))
+;;  structure here is (mostly) {"prismatic/schem" ["0.1.1" "0.2.0" ...]}
+;;  The exceptions are the mvn based artifacts.  There's a ratelimit in place
+;;  for those artifacts so we get the available versions on demand instead.
+(defonce artifacts
+  (atom (if (.exists (io/as-file artifacts-file))
+          (->> artifacts-file slurp edn/read-string (into {}))
+          {})
+        :meta {:last-modified
+               (get-last-modified-from-file artifacts-file)}))
 (def millis-per-day (* 24 60 60 1000))
 
 (defn- get-proxy-opts
@@ -37,7 +46,8 @@
   []
   (or (empty? @artifacts)
       (if-let [last-modified (some-> artifacts meta :last-modified)]
-        (neg? (- millis-per-day (- (.getTime (java.util.Date.)) last-modified)))
+        (neg? (- millis-per-day
+                 (- (.getTime (java.util.Date.)) last-modified)))
         true)))
 
 (defn- edn-read-or-nil
@@ -101,8 +111,11 @@
   [artifact]
   (let [{:keys [body status]} @(http/get (str "https://clojars.org/api/artifacts/"
                                               artifact))]
-    (when (= 200 status)
-      (map :version (:recent_versions (json/read-str body :key-fn keyword))))))
+    (if (= 200 status)
+      (map :version (:recent_versions (json/read-str body :key-fn keyword)))
+      (throw (ex-info (str "Unexpected response from Clojars")
+                      {:status status
+                       :body body})))))
 
 (defn- get-artifacts-from-clojars!
   []
@@ -128,19 +141,88 @@
     (update-artifact-cache!))
   (->> @artifacts keys list*))
 
-(defn artifact-versions
-  "Returns a sorted list of artifact version strings. The list can either come
-  from the artifacts cache, the maven search api or the clojars search api in
-  that order."
-  [{:keys [artifact]}]
-  (->> (or (get @artifacts artifact)
-           (seq (get-mvn-versions! artifact))
-           (get-clojars-versions! artifact))
+(defn- artifact-versions* [artifact-id]
+  (->> (or (get @artifacts artifact-id)
+           (seq (get-mvn-versions! artifact-id))
+           (get-clojars-versions! artifact-id))
        distinct
        versions/version-sort
        reverse
        list*))
 
+(defn artifact-versions
+  "Returns a sorted list of artifact version strings. The list can either come
+  from the artifacts cache, the maven search api or the clojars search api in
+  that order."
+  [{:keys [artifact]}]
+  (artifact-versions* artifact))
+
+(defn- jar-at-the-top-of-dependency-hierarchy [new-deps]
+  ;; We only need to consider the dep at the top of the hierarchy because when we
+  ;; require those namespaces the rest of  the transitive deps will get pulled in
+  ;; too.
+  (letfn [(->jar [^File f]
+            (JarFile. f))]
+    (let [top-level-dep  (-> new-deps keys first)]
+      (-> top-level-dep meta :file ->jar))))
+
+(defn- make-resolve-missing-aware-of-new-deps!
+  "Once the deps are available on cp we still have to load them and
+  reset slamhound's cache to make resolve-missing work."
+  [^JarFile jar]
+  (doseq [new-namespace (find/find-namespaces-in-jarfile jar)]
+    (try
+      (require new-namespace)
+      (catch Exception _
+        ;; I've seen this happen after adding core.async as a dependency.
+        ;; It also happens if you try to require namespaces that no longer work,
+        ;; like compojure.handler.
+        ;; A failure here isn't a big deal, it only means that resolve-missing
+        ;; isn't going to work until the namespace has been loaded manually.
+        )))
+  (slamhound/reset)
+  (slamhound-regrow/clear-cache!))
+
+(defn- parse-coordinates [coordinates-str]
+  (let [coords (try (->> coordinates-str edn/read-string (take 2) vec)
+                    (catch Exception _))]
+    (if (and (= (count coords) 2)
+             (symbol? (first coords))
+             (string? (second coords)))
+      coords
+      (throw (IllegalArgumentException. (str "Malformed dependency vector: "
+                                             coordinates-str))))))
+
+(defn- ensure-coordinates-exist!
+  [[artifact-id artifact-version :as coordinates]]
+  (when (stale-cache?)
+    (update-artifact-cache!))
+  (if-let [versions (artifact-versions* (str artifact-id))]
+    (if ((set versions) artifact-version)
+      coordinates
+      (throw (IllegalArgumentException.
+              (str "Version " artifact-version
+                   " does not exist for " artifact-id
+                   ". Available versions are " (pr-str (vec versions))))))
+    (throw (IllegalArgumentException. (str "Can't find artifact '"
+                                           artifact-id "'")))))
+
+(defn- add-dependencies! [coordinates]
+  ;; Just so we can mock this out during testing
+  (let [repos {"clojars" "https://clojars.org/repo"
+               "central" "https://repo1.maven.org/maven2/"}]
+    (pomegranate/add-dependencies
+     :coordinates [coordinates] :repositories repos)))
+
+(defn- hotload-dependency! [coordinates]
+  (-> (add-dependencies! coordinates)
+      jar-at-the-top-of-dependency-hierarchy
+      make-resolve-missing-aware-of-new-deps!))
+
 (defn hotload-dependency
-  []
-  (throw (IllegalArgumentException. "Temporarily disabled until a solution for java 10 is found.")))
+  [{:keys [coordinates]}]
+  (->> coordinates
+       parse-coordinates
+       ensure-coordinates-exist!
+       (hotload-dependency!)
+       (str/join " ")))